Fixed bug when source files NOT cleaned up
The bug fix was actually just a simple adding of a !
in front of willDelete() so that the logic would actually
be consistent with its comment.  But while I was at
it, I went ahead and made the "allowing" of scanned files
to be a bit more generic and handled earlier in the
process.