Fix bug in single transmission recovery
The case could theoretically occur where all parts of a
bin are received by the server but the connection still
encounters an error.  In that case, the original logic would
simply attempt to send the entire thing again since it would
not accept that the number of parts received could be
equal to the number of parts sent.